Removed warnings for missing variables and signed/unsigned comparisons and added debug option for llvm compiler

+++ hdr/demag.hpp
@@ -36,7 +36,7 @@
 namespace demag{
 
 	extern bool fast;
-	extern int update_rate;
+	extern uint64_t update_rate;
 	
 	extern void init();
 	extern void update();
